Remove parseJson tests, refactor logger and update dependents
This commit removes parseJson related files and updates dependent elements that are impacted by this change. In addition, it modifies log functions to logger and updates all instances accordingly. Other code improvements like the renaming of the RunnerContext to MessageContext in MessageType and index files are also covered in this commit.