feat: Update package.json and add .npmignore file
The package.json file now has the correct syntax for the dist path. In order to specify the files and directories that should be ignored when the package is published to npm, a new .npmignore file has been added. This includes logs, runtime data, coverage data, cache, and many more.