aider: Added a table of contents to the `README.md` file.
# Aider chat conversation:

USER: add a table of contents

ASSISTANT: To add a table of contents to the `README.md` file, I will insert a new section at the beginning of the document, right after the main title and badges. The table of contents will include links to the main sections of the README.

Here is the diff for the changes:

```diff
--- README.md
+++ README.md
@@ ... @@
 # Autonomais

 **Effortlessly simple agent-based workflows!**

+## Table of Contents
+
+- [Why Autonomais?](#why-autonomais)
+- [Getting Started](#getting-started)
+- [Contributing](#contributing)
+- [Technologies Used](#technologies-used)
+- [License](#license)
+
 <hr>

 [![GitHub Repo stars](https://img.shields.io/github/stars/nyvyn/autonomais)](https://github.com/nyvyn/autonomais)
 [![License: MIT](https://img.shields.io/badge/License-MIT-green.svg)](https://opensource.org/licenses/MIT)

 </div>
```

This diff adds a new section for the table of contents with links to the existing sections in the `README.md` file.