Update codebase and refactor tests
Refactored the code base to remove path aliases. Updated example in README and renamed `helloworld.yaml` to `echo.yaml` to reflect the new command in the README. Some tests were also updated to align with the changes made. Paths in the code, jest configurations and tsconfig.json were updated to reflect the removal of path aliases.