refactor(core): update data type of 'lastNode' and its references
Changes have been made in the GraphRunner.ts file. The data type of 'lastNode' has been updated from GraphNode to string. Additionally, all instances where this property is being returned or assigned have also been adjusted to accommodate this change. This focuses on ensuring the right data type is being used and reduces confusion related to the representation of 'lastNode'.