docs(README): update library description and instructions
Extended the introduction to Autonomais in the README and also provided a more detailed description of key concepts including workflows and nodes. Instructions for setting up and running Autonomais as an npm package as well as from the command line have been improved. Various import statements in the example code snippets have been updated for accuracy.